+++ src/components/editor/NoteInlineEditor.tsx
@@ -1,4 +1,5 @@
 import { openUrl } from "@tauri-apps/plugin-opener";
+import { AnimatePresence } from "motion/react";
 import {
 	type MouseEvent as ReactMouseEvent,
 	type ReactNode,
@@ -9,27 +10,12 @@ import {
 	useRef,
 	useState,
 } from "react";
-import {
-	EDITOR_MENU_ACTION_EVENT,
-	type EditorMenuActionDetail,
-} from "../../lib/appEvents";
 import { joinYamlFrontmatter } from "../../lib/notePreview";
-import { invoke } from "../../lib/tauri";
-import { X } from "../Icons";
-import { Button } from "../ui/shadcn/button";
-import {
-	Dialog,
-	DialogContent,
-	DialogDescription,
-	DialogFooter,
-	DialogHeader,
-	DialogTitle,
-} from "../ui/shadcn/dialog";
-import { Input } from "../ui/shadcn/input";
 import { EditorRibbon } from "./EditorRibbon";
 import { ExtractToNoteDialog } from "./ExtractToNoteDialog";
 import { NoteEditorSurface } from "./NoteEditorSurface";
 import { NoteFindBar } from "./NoteFindBar";
+import { NoteLinkDialog, type NoteLinkDialogState } from "./NoteLinkDialog";
 import { NotePropertiesPanel } from "./NotePropertiesPanel";
 import {
 	type SupportedCodeBlockLanguage,
@@ -39,12 +25,12 @@ import {
 import {
 	getMountedEditorContentRoot,
 	getOffsetWithinAncestor,
-	isVisibleEditorHost,
 } from "./hooks/editorDomUtils";
 import { useExtractSelectionToNote } from "./hooks/useExtractSelectionToNote";
 import { useNoteEditor } from "./hooks/useNoteEditor";
 import { useNoteFind } from "./hooks/useNoteFind";
 import { useResetScrollOnChange } from "./hooks/useResetScrollOnChange";
+import { useRibbonCommands } from "./hooks/useRibbonCommands";
 import { useTableInlineControls } from "./hooks/useTableInlineControls";
 import {
 	dispatchMarkdownLinkClick,
@@ -52,8 +38,6 @@ import {
 } from "./markdown/editorEvents";
 import { parseWikiLink } from "./markdown/wikiLinkCodec";
 import type { SelectedCodeBlockState } from "./noteEditorOverlayTypes";
-import { isEditorTextColor } from "./textColors";
-import { isEditorTextHighlight } from "./textHighlights";
 import type { NoteInlineEditorProps } from "./types";
 
 function normalizeBody(markdown: string): string {
@@ -66,13 +50,6 @@ type FrontmatterLinkToken =
 
 const FRONTMATTER_LINK_PATTERN =
 	/!?\[\[[^\]\n]+\]\]|\[[^\]\n]+\]\((?:\\.|[^)\n])+\)|https?:\/\/[^\s<>"')\]]+/g;
-
-let lastFocusedNoteEditorHost: HTMLDivElement | null = null;
-
-interface LinkDialogState {
-	href: string;
-	target: "_self" | "_blank";
-}
 
 function areSelectedCodeBlocksEqual(
 	a: SelectedCodeBlockState | null,
@@ -123,39 +100,6 @@ function extractFrontmatterLinkTokens(text: string): FrontmatterLinkToken[] {
 		tokens.push({ kind: "href", raw, href: raw, start, end });
 	}
 	return tokens;
-}
-
-function normalizeEditorHref(value: string): string {
-	const trimmed = value.trim();
-	if (!trimmed) return "";
-	if (
-		trimmed.startsWith("http://") ||
-		trimmed.startsWith("https://") ||
-		trimmed.startsWith("mailto:") ||
-		trimmed.startsWith("tel:") ||
-		trimmed.startsWith("#") ||
-		trimmed.startsWith("/")
-	) {
-		return trimmed;
-	}
-	return `https://${trimmed}`;
-}
-
-function normalizeCalloutType(type: string): string {
-	return type.toLowerCase() === "warn" ? "warning" : type.toLowerCase();
-}
-
-function createCalloutContent(type: string) {
-	return {
-		type: "blockquote",
-		content: [
-			{
-				type: "paragraph",
-				content: [{ type: "text", text: `[!${normalizeCalloutType(type)}]` }],
-			},
-			{ type: "paragraph" },
-		],
-	};
 }
 
 async function openFrontmatterHref(
@@ -214,7 +158,9 @@ export const NoteInlineEditor = memo(function NoteInlineEditor({
 	const selectedCodeBlockRef = useRef<SelectedCodeBlockState | null>(null);
 	const codeBlockCopyResetTimerRef = useRef<number | null>(null);
 	const [codeBlockCopied, setCodeBlockCopied] = useState(false);
-	const [linkDialog, setLinkDialog] = useState<LinkDialogState | null>(null);
+	const [linkDialog, setLinkDialog] = useState<NoteLinkDialogState | null>(
+		null,
+	);
 	const rawTextareaRef = useRef<HTMLTextAreaElement | null>(null);
 	const previousRelPathRef = useRef(relPath);
 
@@ -284,178 +230,23 @@ export const NoteInlineEditor = memo(function NoteInlineEditor({
 		tiptapHostRef,
 	});
 
-	useEffect(() => {
-		if (!editor || mode !== "rich") return;
-
-		const runEditorAction = (action: string) => {
-			const host = tiptapHostRef.current;
-			if (!host || !isVisibleEditorHost(host)) return;
-			const activeElement = document.activeElement;
-			if (activeElement instanceof HTMLElement) {
-				if (host.contains(activeElement)) {
-					lastFocusedNoteEditorHost = host;
-				} else if (lastFocusedNoteEditorHost !== host) {
-					return;
-				}
-			} else if (lastFocusedNoteEditorHost !== host) {
-				return;
-			}
-			const scrollHost = host.closest(
-				".rfNodeNoteEditorBody",
-			) as HTMLElement | null;
-			const scrollTop = scrollHost?.scrollTop ?? 0;
-			const isReadOnlySafeAction =
-				action === "collapse_all_headings" || action === "expand_all_headings";
-			if (!canEdit && !isReadOnlySafeAction) return;
-			const chain = editor
-				.chain()
-				.focus(null, { scrollIntoView: false })
-				.extendMarkRange("link");
-			const handled = (() => {
-				switch (action) {
-					case "bold":
-						return chain.toggleBold().run();
-					case "italic":
-						return chain.toggleItalic().run();
-					case "underline":
-						return chain.toggleUnderline().run();
-					case "strikethrough":
-						return chain.toggleStrike().run();
-					case "heading_1":
-						return chain.toggleHeading({ level: 1 }).run();
-					case "heading_2":
-						return chain.toggleHeading({ level: 2 }).run();
-					case "heading_3":
-						return chain.toggleHeading({ level: 3 }).run();
-					case "collapse_all_headings":
-						return chain.collapseAllHeadings().run();
-					case "expand_all_headings":
-						return chain.expandAllHeadings().run();
-					case "bullet_list":
-						return chain.toggleBulletList().run();
-					case "numbered_list":
-						return chain.toggleOrderedList().run();
-					case "todo_list":
-						return chain.toggleTaskList().run();
-					case "quote":
-						return chain.toggleBlockquote().run();
-					case "code_block":
-						return chain.toggleCodeBlock().run();
-					case "mermaid_chart":
-						return chain
-							.insertContent({
-								type: "codeBlock",
-								attrs: { language: "mermaid" },
-								content: [
-									{
-										type: "text",
-										text: "flowchart TD\n  A[Start] --> B[End]",
-									},
-								],
-							})
-							.run();
-					case "table":
-						return chain
-							.insertTable({ rows: 3, cols: 3, withHeaderRow: true })
-							.run();
-					case "divider":
-						return chain.setHorizontalRule().run();
-					case "extract_selection_to_note":
-						extractToNote.openExtractDialog();
-						return true;
-					case "callout_info":
-						return chain.insertContent(createCalloutContent("info")).run();
-					case "callout_warning":
-						return chain.insertContent(createCalloutContent("warning")).run();
-					case "callout_error":
-						return chain.insertContent(createCalloutContent("error")).run();
-					case "callout_success":
-						return chain.insertContent(createCalloutContent("success")).run();
-					case "callout_tip":
-						return chain.insertContent(createCalloutContent("tip")).run();
-					case "link_set": {
-						const linkAttrs = editor.getAttributes("link") as {
-							href?: string;
-							target?: string;
-						};
-						setLinkDialog({
-							href: linkAttrs.href ?? "",
-							target: linkAttrs.target === "_blank" ? "_blank" : "_self",
-						});
-						return true;
-					}
-					case "link_clear":
-						return chain.unsetLink().run();
-					case "color_clear":
-						return chain.unsetTextColor().run();
-					case "highlight_clear":
-						return chain.unsetTextHighlight().run();
-					default: {
-						if (action.startsWith("color_")) {
-							const color = action.slice("color_".length);
-							if (isEditorTextColor(color)) {
-								return chain.setTextColor(color).run();
-							}
-							return false;
-						}
-						if (action.startsWith("highlight_")) {
-							const highlight = action.slice("highlight_".length);
-							if (isEditorTextHighlight(highlight)) {
-								return chain.setTextHighlight(highlight).run();
-							}
-							return false;
-						}
-						return false;
-					}
-				}
-			})();
-			if (!handled) return;
-			if (scrollHost) {
-				requestAnimationFrame(() => {
-					scrollHost.scrollTop = scrollTop;
-				});
-			}
-		};
-
-		const onEditorMenuAction = (event: Event) => {
-			const detail =
-				event instanceof CustomEvent
-					? (event.detail as EditorMenuActionDetail | null)
-					: null;
-			if (!detail?.action) return;
-			runEditorAction(detail.action);
-		};
-
-		window.addEventListener(EDITOR_MENU_ACTION_EVENT, onEditorMenuAction);
-		return () => {
-			window.removeEventListener(EDITOR_MENU_ACTION_EVENT, onEditorMenuAction);
-		};
-	}, [canEdit, editor, extractToNote.openExtractDialog, mode]);
-
-	useEffect(() => {
-		if (!onRegisterCalloutInserter) return;
-		if (!editor || mode !== "rich") {
-			onRegisterCalloutInserter(null);
-			return;
-		}
-		onRegisterCalloutInserter((type: string) => {
-			const host = tiptapHostRef.current?.closest(
-				".rfNodeNoteEditorBody",
-			) as HTMLElement | null;
-			const scrollTop = host?.scrollTop ?? 0;
-			editor
-				.chain()
-				.focus(null, { scrollIntoView: false })
-				.insertContent(createCalloutContent(type))
-				.run();
-			if (host) {
-				requestAnimationFrame(() => {
-					host.scrollTop = scrollTop;
-				});
-			}
-		});
-		return () => onRegisterCalloutInserter(null);
-	}, [editor, mode, onRegisterCalloutInserter]);
+	useRibbonCommands({
+		editor,
+		canEdit,
+		mode,
+		tiptapHostRef,
+		tiptapHostNode,
+		onOpenLinkDialog: useCallback(
+			(href: string, target: "_self" | "_blank") => {
+				setLinkDialog({ href, target });
+			},
+			[],
+		),
+		onTriggerExtractToNote: extractToNote.canExtractToNote
+			? extractToNote.openExtractDialog
+			: undefined,
+		onRegisterCalloutInserter,
+	});
 
 	const handleFrontmatterChange = (next: string | null) => {
 		const normalizedFrontmatter = next?.trim().length ? next : null;
@@ -726,76 +517,10 @@ export const NoteInlineEditor = memo(function NoteInlineEditor({
 		return () => observer.disconnect();
 	}, [editor, mode]);
 
-	useEffect(() => {
-		const host = tiptapHostNode;
-		if (!host) return;
-		const handleFocusIn = () => {
-			lastFocusedNoteEditorHost = host;
-		};
-		const handleFocusOut = () => {
-			const currentHost = host;
-			window.setTimeout(() => {
-				if (
-					lastFocusedNoteEditorHost === currentHost &&
-					!currentHost.contains(document.activeElement)
-				) {
-					lastFocusedNoteEditorHost = null;
-				}
-			}, 0);
-		};
-		handleFocusOut();
-		host.addEventListener("focusin", handleFocusIn);
-		host.addEventListener("focusout", handleFocusOut);
-		return () => {
-			host.removeEventListener("focusin", handleFocusIn);
-			host.removeEventListener("focusout", handleFocusOut);
-			if (lastFocusedNoteEditorHost === host) {
-				lastFocusedNoteEditorHost = null;
-			}
-		};
-	}, [tiptapHostNode]);
-
 	const handleTiptapHostRef = useCallback((node: HTMLDivElement | null) => {
 		tiptapHostRef.current = node;
 		setTiptapHostNode(node);
 	}, []);
-
-	const closeLinkDialog = useCallback(() => {
-		setLinkDialog(null);
-	}, []);
-
-	const applyLinkDialog = useCallback(() => {
-		if (!editor || !canEdit || !linkDialog) return;
-		const href = normalizeEditorHref(linkDialog.href);
-		const chain = editor
-			.chain()
-			.focus(null, { scrollIntoView: false })
-			.extendMarkRange("link");
-		if (!href) {
-			chain.unsetLink().run();
-			setLinkDialog(null);
-			return;
-		}
-		chain
-			.setLink({
-				href,
-				target: linkDialog.target,
-				rel: linkDialog.target === "_blank" ? "noopener noreferrer" : undefined,
-			})
-			.run();
-		setLinkDialog(null);
-	}, [canEdit, editor, linkDialog]);
-
-	const removeLinkFromDialog = useCallback(() => {
-		if (!editor || !canEdit) return;
-		editor
-			.chain()
-			.focus(null, { scrollIntoView: false })
-			.extendMarkRange("link")
-			.unsetLink()
-			.run();
-		setLinkDialog(null);
-	}, [canEdit, editor]);
 
 	const copySelectedCodeBlock = useCallback(() => {
 		if (!selectedCodeBlock) return;
@@ -922,18 +647,20 @@ export const NoteInlineEditor = memo(function NoteInlineEditor({
 					/>
 				) : null}
 			</div>
-			{canEdit && editor ? (
-				<EditorRibbon
-					editor={editor}
-					canEdit={canEdit}
-					className="rfNodeNoteEditorRibbonBottom"
-					onExtractSelectionToNote={
-						extractToNote.canExtractToNote
-							? extractToNote.openExtractDialog
-							: undefined
-					}
-				/>
-			) : null}
+			<AnimatePresence>
+				{canEdit && editor ? (
+					<EditorRibbon
+						editor={editor}
+						canEdit={canEdit}
+						className="rfNodeNoteEditorRibbonBottom"
+						onExtractSelectionToNote={
+							extractToNote.canExtractToNote
+								? extractToNote.openExtractDialog
+								: undefined
+						}
+					/>
+				) : null}
+			</AnimatePresence>
 			<ExtractToNoteDialog
 				state={extractToNote.dialogState}
 				onClose={extractToNote.closeExtractDialog}
@@ -941,79 +668,12 @@ export const NoteInlineEditor = memo(function NoteInlineEditor({
 				onTitleChange={extractToNote.setExtractTitle}
 				onDestinationDirChange={extractToNote.setExtractDestinationDir}
 			/>
-			<Dialog
-				open={linkDialog !== null}
-				onOpenChange={(open) => {
-					if (!open) closeLinkDialog();
-				}}
-			>
-				<DialogContent
-					className="editorLinkDialog"
-					onOpenAutoFocus={(event) => {
-						const input = document.querySelector<HTMLInputElement>(
-							".editorLinkDialogInput",
-						);
-						if (!input) return;
-						event.preventDefault();
-						input.focus();
-						input.select();
-					}}
-				>
-					<DialogHeader>
-						<DialogTitle>Link</DialogTitle>
-						<DialogDescription>
-							Paste a URL, or leave it blank to remove the link.
-						</DialogDescription>
-					</DialogHeader>
-					<form
-						className="editorLinkDialogForm"
-						onSubmit={(event) => {
-							event.preventDefault();
-							applyLinkDialog();
-						}}
-					>
-						<Input
-							className="editorLinkDialogInput"
-							value={linkDialog?.href ?? ""}
-							onChange={(event) =>
-								setLinkDialog((current) =>
-									current ? { ...current, href: event.target.value } : current,
-								)
-							}
-							placeholder="https://example.com"
-							aria-label="Link URL"
-						/>
-						<label className="editorLinkDialogCheckbox">
-							<input
-								type="checkbox"
-								checked={linkDialog?.target === "_blank"}
-								onChange={(event) =>
-									setLinkDialog((current) =>
-										current
-											? {
-													...current,
-													target: event.target.checked ? "_blank" : "_self",
-												}
-											: current,
-									)
-								}
-							/>
-							<span>Open in new tab</span>
-						</label>
-						<DialogFooter className="editorLinkDialogActions">
-							<Button
-								type="button"
-								variant="ghost"
-								onClick={removeLinkFromDialog}
-							>
-								<X size="var(--icon-md)" />
-								Remove
-							</Button>
-							<Button type="submit">Apply</Button>
-						</DialogFooter>
-					</form>
-				</DialogContent>
-			</Dialog>
+			<NoteLinkDialog
+				editor={editor}
+				canEdit={canEdit}
+				state={linkDialog}
+				onStateChange={setLinkDialog}
+			/>
 		</div>
 	);
 });
